Revival commit
Emacs interface for Guix is about to be separated from the Guix code, so
this commit removes the old code and adds the according code and
data (elisp, scheme and texi documentation) to this repository.

The code itself wasn't touched (only header commentaries were changed),
so all this stuff is not in a working state yet.