scheme: Remove "helper.scm.in"
"helper.scm" was used to add required directories to Guile load-path, to
remember the value of "guix pull"-ed directory (to pass it to elisp
side), and to load "main.scm".  This can be done on elisp side.

* scheme/emacs-guix/helper.scm.in: Remove.
* elisp/guix-backend.el (guix-latest-directory): New procedure.
(guix-helper-file, guix-set-directory): Remove.
(guix-after-start-repl-hook): Remove it.
(guix-get-guile-program): Rename to...
(guix-repl-guile-program): ... this.  Adjust to return the required path.
(guix-start-repl-maybe): Adjust accordingly.
(guix-start-repl): Do not add module directory here.