elisp/ui-package: Remove 'output-info' interface
It does not provide anything useful comparing with the 'package-info'
interface; and it was added just for "symmetry": since there were both
'package-list' and 'output-list', I thought it would be also good to add
'output-info', but it is really useless.

* elisp/guix-ui-package.el: Remove all the 'output-info' stuff.
(guix-package-info-type): Remove custom variable.
(guix-package-type): Rename to...
(guix-package-list-type): ... this.
* elisp/guix-base.el (guix-update-buffers-maybe-after-pull): Do not
check for 'guix-output-info-mode'.
* elisp/guix-ui.el (guix-ui-buffer?): Likewise.
* doc/emacs-guix.texi (Configuration): Remove 'output-info' buffer name
and keymap.