elisp/about: Factorize Help/About buffer displaying
A very similar code is used to insert the content and to display "Help"
and "About" buffers.  This commit generalize these commonalities.

* elisp/guix-about.el (guix-help-revert, guix-about-revert): Remove,
replace with...
(guix-help-make-revert-function): ... this.  New procedure.
(guix-help-insert-content, guix-about-insert-content): Do not erase the
buffer, as it is done by...
(guix-help-reinsert-content): ... this.  New procedure.
(guix-help, guix-about): Do not set up a target buffer as it is done by...
(guix-help-display-buffer): ... this. New procedure.