elisp/about: Display Help/About buffers in the current window
Using 'with-output-to-temp-buffer' macro has some downsides.
Particularly, by default, it displays a buffer in another window; also
it does not allow to move the point position.  So add a major mode, and
display a buffer in the current window using it.

* elisp/guix-about.el (guix-help-mode-map): New variable.
(guix-help-mode): New major mode.
(guix-help-display-buffer): Use it and switch to the resulting buffer.