elisp/ui-generation: Handle system profile specially if it is specified
System profiles have more things to show than usual (non-system)
profiles, so if a system profile is specified to the common generation
commands, it should be handled specially.

* elisp/guix-ui-generation.el (guix-generations)
(guix-last-generations, guix-generations-by-time): Check if profile is a
system one, and call the according system commands if needed.
* elisp/guix-ui-profile.el (guix-profile-list-show-generations): Do not
check for system profile as it is checked in 'guix-generations'.