Rename 'location' to 'location-file' where needed
This is a followup to commit d3881fec140de868265a28376311c9750efdd351.

There is a distinction between these terms as 'location-file' is just a
file, and 'location' is a file plus a position in that file.
So (ideally) there should be the same distinction in the command names,
however most commands already have long names, and using
'guix-packages-by-location-file' instead of 'guix-packages-by-location'
will probably add nothing to the meaning of this command and 5
characters to its name.  So this command is left unchanged, but such
commands as 'guix-find-package-location' and
'guix-find-package-definition' may bring confusion.  Hopefully, using
'guix-find-package-location-file' name will make it more clear.

* elisp/guix-read.el (guix-read-package-location): Rename to...
(guix-read-package-location-file): ... this.
* elisp/guix-package.el (guix-find-package-location): Rename to...
(guix-find-package-location-file): ... this.
* elisp/guix-help.el (guix-help-specifications): Adjust accordingly.
* elisp/guix-ui-package.el (guix-packages-by-location): Likewise.
* doc/emacs-guix.texi (Package Locations): Likewise.