Use (id spec) list instead of just spec for package inputs
Previously, only package specifications were passed to the elisp side
for package inputs.  But a package is not uniquely defined by its name
and version, so when you pressed a package input button in "Guix Package
Info" buffer, a package with this specification would be displayed, but
it might not be the real input (for example, 'ld-wrapper' package has
'guile@2.2.3' input, but it is not the default 'guile', it is
'guile-final' package from (gnu packages commencement) module, which has
the same name and version).

So now, (id spec) list is passed to the elisp side instead.  This allows
to define the exact package input, not relying only on its specification.

* scheme/emacs-guix/packages.scm (package-inputs-names): Rename to...
(package-inputs-sexps): ... this.  Return (address spec) list.
(%package-param-alist): Adjust accordingly.
(id->package-pattern, id->output-pattern): Adjust to handle different
kinds of the argument.  So 'id' search type can be either a number (id),
a string (specification) or (id spec) list.

* elisp/guix-ui-package.el (guix-package-info-insert-name-buttons)
(guix-package-info-insert-name-button): New procedures.