elisp: Set 'guix-directory' when needed
Before the previous commit (6476f612aa6c71779abf2be6d887bec3ac89ce00),
'guix-directory' variable was blindly set to the latest "guix pull"
directory.  But it is not good because a user may not have that pulled
directory.  It is better to ask Guile where guix modules are placed and
just use this directory.

* elisp/guix-repl.el (guix-directory): New function that sets
'guix-directory' variable if needed.
(guix-read-directory): Use 'guix-directory' function instead of variable.
* elisp/guix-location.el (guix-find-location): Likewise.
* elisp/guix-license.el (guix-license-file): Likewise.
* doc/emacs-guix.texi (More Package Commands): Remove documentation of
'guix-directory'.